pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ab-reviewer"
version = "0.1.0"
description = "Automated code review workflow with AI integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "AB Code Reviewer", email = "ab@example.com"},
]
keywords = ["code-review", "ai", "quality", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "colorama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "pylint>=2.15.0",
    "bandit>=1.7.0",
]

[project.scripts]
ab-reviewer = "ab_reviewer.cli:main"

[project.urls]
Homepage = "https://github.com/ab-code-reviewer/ab-reviewer"
Repository = "https://github.com/ab-code-reviewer/ab-reviewer"
Issues = "https://github.com/ab-code-reviewer/ab-reviewer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ab_reviewer*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pylint]
max-line-length = 88
disable = ["C0114", "C0116"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (require external tools)",
    "slow: marks tests as slow running",
    "external: marks tests that require external services",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-m", "not integration",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "click.*",
    "yaml.*",
    "colorama.*",
]
ignore_missing_imports = true
